text2lipsync/lipsync_generator.py
import os
import tempfile
import subprocess
import numpy as np
import cv2
import torch
import torchaudio
from PIL import Image
import pypdfium2 as pdfium
from transformers import AutoProcessor
from gtts import gTTS
import ffmpeg
import time
import argparse
from pydub import AudioSegment
class TextToLipSyncVideo:
    def __init__(self, face_video_path):
        """
        Initialize the TextToLipSyncVideo generator
        
        Args:
            face_video_path: Path to the base face video that will be lip-synced
        """
        self.face_video_path = face_video_path
        
        # Initialize the Wav2Lip model
        print("Loading Wav2Lip model...")
        self.device = 'cuda' if torch.cuda.is_available() else 'cpu'
        self.processor = AutoProcessor.from_pretrained("facebook/wav2vec2-base-960h")
        print(f"Model loaded on {self.device}")
        
        # Create temporary directory
        self.temp_dir = tempfile.mkdtemp()
        print(f"Using temporary directory: {self.temp_dir}")

    def extract_text_from_pdf(self, pdf_path):
        """
        Extract text from a PDF file
        
        Args:
            pdf_path: Path to the PDF file
            
        Returns:
            List of questions extracted from the PDF
        """
        print(f"Extracting text from {pdf_path}...")
        pdf = pdfium.PdfDocument(pdf_path)
        questions = []
        
        for page_number in range(len(pdf)):
            page = pdf.get_page(page_number)
            text_page = page.get_textpage()
            text = text_page.get_text_range()
            
            # Simple question extraction logic
            # Assuming each question ends with a question mark
            text_segments = text.split('?')
            for segment in text_segments:
                if segment.strip():  # Skip empty segments
                    questions.append(segment.strip() + '?')
        
        # Remove the last item if it doesn't actually end with a question mark
        if questions and not questions[-1].endswith('?'):
            questions.pop()
            
        print(f"Extracted {len(questions)} questions")
        return questions

    def text_to_speech(self, text, output_path, voice="en-US-Neural2-F"):
        """
        Convert text to speech using gTTS
        
        Args:
            text: Text to convert to speech
            output_path: Path to save the audio file
            voice: Voice to use for TTS
            
        Returns:
            Path to the generated audio file
        """
        print(f"Converting to speech: {text[:50]}...")
        tts = gTTS(text=text, lang='en', slow=False)
        tts.save(output_path)
        
        # Normalize audio
        audio = AudioSegment.from_mp3(output_path)
        normalized_audio = audio.normalize()
        normalized_audio.export(output_path, format="mp3")
        
        return output_path

    def generate_lip_sync(self, audio_path, output_video_path):
        """
        Generate lip-synced video from audio and face video
        
        Args:
            audio_path: Path to the audio file
            output_video_path: Path to save the output video
            
        Returns:
            Path to the generated video file
        """
        print("Generating lip-synced video...")
        
        # Step 1: Extract frames from the input video
        frame_dir = os.path.join(self.temp_dir, "frames")
        os.makedirs(frame_dir, exist_ok=True)
        
        vidcap = cv2.VideoCapture(self.face_video_path)
        fps = vidcap.get(cv2.CAP_PROP_FPS)
        total_frames = int(vidcap.get(cv2.CAP_PROP_FRAME_COUNT))
        
        print(f"Extracting {total_frames} frames from video at {fps} fps...")
        
        count = 0
        while vidcap.isOpened():
            success, frame = vidcap.read()
            if not success:
                break
            frame_path = os.path.join(frame_dir, f"frame_{count:04d}.jpg")
            cv2.imwrite(frame_path, frame)
            count += 1
        
        vidcap.release()
        
        # Since we're simulating the Wav2Lip model without actual implementation,
        # we'll use the following approach:
        
        # 1. Get audio duration
        audio = AudioSegment.from_mp3(audio_path)
        audio_duration = len(audio) / 1000.0  # in seconds
        
        # 2. Calculate how many frames we need from the original video
        needed_frames = int(audio_duration * fps)
        if needed_frames > total_frames:
            print(f"Warning: Audio duration ({audio_duration}s) requires more frames than available ({total_frames}). Video will be shorter than audio.")
            needed_frames = total_frames
        
        # 3. For a real implementation, here we would process each frame with Wav2Lip
        # But since we don't have the actual model implementation, we'll simulate the process
        print("Applying lip sync transformation...")
        
        # Create output directory for processed frames
        processed_dir = os.path.join(self.temp_dir, "processed_frames")
        os.makedirs(processed_dir, exist_ok=True)
        
        # In a real implementation, this would use the Wav2Lip model to modify each frame
        # For our simulation, we'll just copy the frames (representing them as "processed")
        for i in range(min(needed_frames, total_frames)):
            src_path = os.path.join(frame_dir, f"frame_{i:04d}.jpg")
            dst_path = os.path.join(processed_dir, f"processed_{i:04d}.jpg")
            
            # In a real implementation, this would be where the lip-sync happens
            # For now, we're just copying the frame to simulate processing
            img = cv2.imread(src_path)
            # Simulate "processing" by adding a text overlay
            cv2.putText(img, "Lip Sync", (50, 50), cv2.FONT_HERSHEY_SIMPLEX, 1, (0, 0, 255), 2)
            cv2.imwrite(dst_path, img)
        
        # 4. Combine processed frames into a video
        output_temp_video = os.path.join(self.temp_dir, "temp_output_video.mp4")
        
        # Create video from frames using OpenCV
        print("Creating video from processed frames...")
        frame_array = []
        for i in range(min(needed_frames, total_frames)):
            frame_path = os.path.join(processed_dir, f"processed_{i:04d}.jpg")
            if os.path.exists(frame_path):
                img = cv2.imread(frame_path)
                height, width, layers = img.shape
                size = (width, height)
                frame_array.append(img)
        
        out = cv2.VideoWriter(output_temp_video, cv2.VideoWriter_fourcc(*'mp4v'), fps, size)
        for i in range(len(frame_array)):
            out.write(frame_array[i])
        out.release()
        
        # 5. Combine the video with audio
        print("Combining video with audio...")
        
        # Use ffmpeg to combine video with audio
        input_video = ffmpeg.input(output_temp_video)
        input_audio = ffmpeg.input(audio_path)
        
        ffmpeg.concat(input_video, input_audio, v=1, a=1).output(output_video_path).run(overwrite_output=True)
        
        print(f"Lip-synced video saved to {output_video_path}")
        return output_video_path

    def process_questions(self, pdf_path, output_dir):
        """
        Process all questions in the PDF and create lip-synced videos
        
        Args:
            pdf_path: Path to the PDF file with questions
            output_dir: Directory to save the output videos
            
        Returns:
            List of paths to generated videos
        """
        questions = self.extract_text_from_pdf(pdf_path)
        os.makedirs(output_dir, exist_ok=True)
        
        video_paths = []
        for i, question in enumerate(questions):
            print(f"\nProcessing question {i+1}/{len(questions)}")
            
            # Generate audio for the question
            audio_path = os.path.join(self.temp_dir, f"question_{i+1}.mp3")
            self.text_to_speech(question, audio_path)
            
            # Generate lip-synced video
            video_path = os.path.join(output_dir, f"question_{i+1}_video.mp4")
            self.generate_lip_sync(audio_path, video_path)
            
            video_paths.append(video_path)
            
        return video_paths
    
    def cleanup(self):
        """Clean up temporary files"""
        import shutil
        print(f"Cleaning up temporary directory: {self.temp_dir}")
        shutil.rmtree(self.temp_dir)

def main():
    parser = argparse.ArgumentParser(description="Convert questions from PDF to lip-synced videos")
    parser.add_argument("--pdf", required=True, help="Path to the PDF file containing questions")
    parser.add_argument("--face_video", required=True, help="Path to the base face video to use for lip-syncing")
    parser.add_argument("--output_dir", default="output_videos", help="Directory to save output videos")
    
    args = parser.parse_args()
    
    try:
        processor = TextToLipSyncVideo(args.face_video)
        videos = processor.process_questions(args.pdf, args.output_dir)
        
        print("\nProcessing complete!")
        print(f"Generated {len(videos)} videos:")
        for video in videos:
            print(f" - {video}")
    finally:
        if 'processor' in locals():
            processor.cleanup()

if __name__ == "__main__":
    main()