text2lipsync/requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
torch>=1.7.0
torchaudio>=0.7.0
transformers>=4.15.0
pypdfium2>=2.0.0
gtts>=2.2.0
ffmpeg-python>=0.2.0
pydub>=0.25.0
Pillow>=8.0.0